requirements_viz.txt
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
docker-pycreds==0.4.0
fonttools==4.54.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.0.2
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.14.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.1
